requirements.txt
accelerate==1.0.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
argcomplete==3.5.1
async-timeout==4.0.3
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ckip-transformers==0.3.4
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==43.0.1
datamodel-code-generator==0.26.1
datasets==3.0.1
deepdiff==8.0.1
dill==0.3.8
dnspython==2.7.0
email_validator==2.2.0
faiss-cpu==1.9.0
filelock==3.16.1
filetype==1.2.0
FlagEmbedding==1.2.11
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gensim==4.3.3
genson==1.3.0
hmmlearn==0.3.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
inflect==5.6.2
isort==5.13.2
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
onnxruntime==1.19.2
opencc-python-reimplemented==0.1.7
opencv-python==4.10.0.84
orderly-set==5.2.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
peft==0.13.2
pillow==10.4.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.11
PyPDF2==3.0.1
pypdfium2==4.30.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
rapidocr-onnxruntime==1.3.24
rapidocr-paddle==1.3.27
rapidocr-pdf==0.1.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.1.1
shapely==2.0.6
six==1.16.0
smart-open==7.0.5
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.2
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
TorchCRF==1.1.0
torchdata==0.7.1
torchtext==0.17.0
torchvision==0.19.1+cu118
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
xxhash==3.5.0
yarl==1.15.5
